+++ apps/3000-home/next.config.js
@@ -1,6 +1,10 @@
 const { withNx } = require('@nrwl/next/plugins/with-nx');
 
 const NextFederationPlugin = require('@module-federation/nextjs-mf');
+// const {
+//   promiseTemplate,
+//   promiseFactory,
+// } = require('@module-federation/nextjs-mf/utils');
 
 /**
  * @type {import('@nrwl/next/plugins/with-nx').WithNxOptions}
@@ -14,17 +18,22 @@ const nextConfig = {
   webpack(config, options) {
     const { isServer } = options;
     if (!options.isServer) {
+      const remotes = {
+        shop: `shop@http://localhost:3001/_next/static/${
+          isServer ? 'ssr' : 'chunks'
+        }/remoteEntry.js`,
+        checkout: `checkout@http://localhost:3002/_next/static/${
+          isServer ? 'ssr' : 'chunks'
+        }/remoteEntry.js`,
+      };
+
       config.plugins.push(
         new NextFederationPlugin({
           name: 'home_app',
           filename: 'static/chunks/remoteEntry.js',
           remotes: {
-            shop: `shop@http://localhost:3001/_next/static/${
-              isServer ? 'ssr' : 'chunks'
-            }/remoteEntry.js`,
-            checkout: `checkout@http://localhost:3002/_next/static/${
-              isServer ? 'ssr' : 'chunks'
-            }/remoteEntry.js`,
+            shop: remotes.shop,
+            checkout: remotes.checkout,
           },
           exposes: {
             './SharedNav': './components/SharedNav',
