+++ package.json
@@ -8,6 +8,8 @@
   },
   "private": true,
   "dependencies": {
+    "@module-federation/nextjs-mf": "link:./dist/packages/nextjs-mf",
+    "@module-federation/node": "link:./dist/packages/node",
     "@swc/helpers": "~0.3.3",
     "core-js": "^3.6.5",
     "fast-glob": "^3.2.12",
@@ -16,10 +18,7 @@
     "react-dom": "18.2.0",
     "regenerator-runtime": "0.13.7",
     "tslib": "^2.3.0",
-    "webpack": "^5.74.0",
-    "@module-federation/nextjs-mf": "link:./dist/packages/nextjs-mf",
-    "@module-federation/node": "link:./dist/packages/node",
-    "@internal/shared": "link:./packages/shared"
+    "webpack": "^5.74.0"
   },
   "devDependencies": {
     "@nrwl/cli": "14.7.3",
@@ -32,7 +31,6 @@
     "@nrwl/web": "14.7.5",
     "@nrwl/workspace": "14.7.3",
     "@swc/cli": "~0.1.55",
-    "@swc/core": "^1.2.173",
     "@swc/jest": "0.2.20",
     "@testing-library/react": "13.3.0",
     "@types/jest": "28.1.1",
