+++ packages/nextjs-mf/.eslintrc.json
@@ -4,7 +4,9 @@
   "overrides": [
     {
       "files": ["*.ts", "*.tsx", "*.js", "*.jsx"],
-      "rules": {}
+      "rules": {
+        "@typescript-eslint/no-var-requires": false
+      }
     },
     {
       "files": ["*.ts", "*.tsx"],
